Allow configurable vendor path for Composer packages
The current install steps assume Composer is run inside the plugin folder and installs packages in /Plugin/CakeResque/vendor. This change allows you to also install the CakeResque plugin with Composer. The plugin and all its dependencies are installed in the default Composer location (/Vendor for CakePHP). The default behavior stays the same as it was before.